+++ setuptools_rust/build.py
@@ -39,6 +39,7 @@ class build_rust(RustCommand):
             "t",
             "directory for temporary files (cargo 'target' directory) ",
         ),
+        ("target", None, "Build for the target triple"),
     ]
     boolean_options = ["inplace", "debug", "release", "qbuild"]
 
@@ -50,6 +51,7 @@ class build_rust(RustCommand):
         self.qbuild = None
         self.build_temp = None
         self.plat_name = None
+        self.target = os.getenv("CARGO_BUILD_TARGET")
 
     def finalize_options(self):
         super().finalize_options()
@@ -66,9 +68,9 @@ class build_rust(RustCommand):
         # If we are on a 64-bit machine, but running a 32-bit Python, then
         # we'll target a 32-bit Rust build.
         # Automatic target detection can be overridden via the CARGO_BUILD_TARGET
-        # environment variable.
-        if os.getenv("CARGO_BUILD_TARGET"):
-            return os.environ["CARGO_BUILD_TARGET"]
+        # environment variable or --target command line option
+        if self.target:
+            return self.target
         elif self.plat_name == "win32":
             return "i686-pc-windows-msvc"
         elif self.plat_name == "win-amd64":
@@ -166,7 +168,9 @@ class build_rust(RustCommand):
             if quiet:
                 args.append("-q")
             elif self.verbose:
-                args.append("--verbose")
+                # cargo only have -vv
+                verbose_level = 'v' * min(self.verbose, 2)
+                args.append(f"-{verbose_level}")
 
         else:
             args = (
@@ -180,7 +184,9 @@ class build_rust(RustCommand):
             if quiet:
                 args.append("-q")
             elif self.verbose:
-                args.append("--verbose")
+                # cargo only have -vv
+                verbose_level = 'v' * min(self.verbose, 2)
+                args.append(f"-{verbose_level}")
 
             args.extend(["--", "--crate-type", "cdylib"])
             args.extend(ext.rustc_flags or [])